+++ app/Providers/Filament/AdminPanelProvider.php
@@ -3,6 +3,8 @@
 namespace App\Providers\Filament;
 
 use App\Filament\Pages\Auth\Login;
+use App\Filament\Widgets\BotTierDistributionWidget;
+use App\Filament\Widgets\InFlightWithdrawalsWidget;
 use Filament\Http\Middleware\Authenticate;
 use Filament\Http\Middleware\AuthenticateSession;
 use Filament\Http\Middleware\DisableBladeIconComponents;
@@ -34,6 +36,8 @@ class AdminPanelProvider extends PanelProvider
             ->discoverWidgets(in: app_path('Filament/Widgets'), for: 'App\\Filament\\Widgets')
             ->widgets([
                 Widgets\AccountWidget::class,
+                InFlightWithdrawalsWidget::class,
+                BotTierDistributionWidget::class,
             ])
             ->middleware([
                 EncryptCookies::class,